requirements.txt
Flask
requests
python-dotenv
flask-cors
openai
